Add Logarithmic function and parameter.
Using the 'Fct' class to draw the function.
Using class 'Plot' to contain all the static functions that are
to be plotted. This class also contains static variables that can
be altered from the main loop and used to adjust the 'parameters'
inside of the static functions (Other than 'x' which is passed
by copy in the function signature).

Widget behaviour:
Toggle hides / shows.
Parameter 'b' controls the base of the logarithmic function
being plotted.