Refactor CMakeLists.txt to use a more procedural approach
The rest of the chapters will also be updated to use the same setup.

Targets and directories are stored in lists.
Combined, these serve as the single source of truth for both
generating messages and selecting the correct build target.

Also, an option for building all targets at once was added,
currently this is done by setting SELECTED_TARGET to "ALL".