[ch10] CMakeLists.txt: Copy data folder to dir. of executable
The previous use of 'file(COPY <path> DESTINATION)' was replaced
with a post-build action to ensure the 'data' files are accessible
regardless of where CMake ends up placing the executable.

Also add an exercise specific suffix to the data folder to avoid
potential name clashes when executables end up in the same folder.